UTOPIA-430 testing ec2 bastion host rds

+++ terraform/rds/mysql_starter_script.sh
@@ -13,7 +13,8 @@ if [[ $ENVIRONMENT=='prod' ]]
 then
     echo 'prod'
     sudo shutdown now -h
-else
+elif [[ $ENVIRONMENT=='dev']]
+then
     echo 'dev'
     poweroff
 fi
